refactor: Moved alot of code so that services layer is used and endpoint is mainly just for authentication and thin
